update set-Permissionslevel, issue principaltype
Fixed the issue when you try to add a group_name, in the body of the rest api the acl is always set on user_name.
That gives the following error when you try to add group names
 `Invoke-RestMethod : {"error_code":"BAD_REQUEST","message":"Principal: UserName(groupname) does not exist"}`

So I've added the option PrincipalType to the parameters and the switch will change the body.

This is not perfect :
-- it could be set to default user_name
-- now you have the parameter Principal for PrincipalName which is little bit confusing
-- Maybe you can let the function capture the all the usernames , and groupnames and decide which type it is.

But my fix is a patch which will handle the issue of the error. As I use this beautiful scripts at work at the moment and I really need a fix I didn't had the time to solve the above.